move middleware to a subpackage of router.
